requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
pydantic-settings
sqlalchemy
chromadb
celery
redis